requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Streamlit for frontend
streamlit==1.28.1

# Data manipulation
numpy==1.26.3
pandas==2.2.0

# HTTP requests
requests==2.31.0

# Data visualization
plotly==5.17.0

# Development and utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4